BuildingMap/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.0.1
pandas==2.1.1
selenium==4.15.0
webdriver-manager==4.0.1 